% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{bundle}
\alias{bundle}
\title{Bundle}
\usage{
bundle(src = "srcpkg", dest = "R")
}
\arguments{
\item{src}{Source directory.}

\item{dest}{Destination directory.}
}
\description{
Bundles R files from nested directory into a flat one.
}
