% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verbose.R
\name{collapse_msg}
\alias{collapse_msg}
\title{Collapse message for CLI}
\usage{
collapse_msg(...)
}
\arguments{
\item{...}{Any number of strings.}
}
\description{
Collapse a message to pass to \code{cli} functions.
}
\keyword{internal}
