% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir_if_not_exists}
\alias{create_dir_if_not_exists}
\title{Create Directory}
\usage{
create_dir_if_not_exists(dir)
}
\arguments{
\item{dir}{Directory to create.}
}
\description{
Create directory if it does not exist.
}
\keyword{internal}
