% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools-document.R, R/devtools-load-all.R
\name{document}
\alias{document}
\title{Bundle then document}
\usage{
document(..., bundle_opts = list())

document(..., bundle_opts = list())
}
\arguments{
\item{...}{Arguments passed to \code{\link[devtools:load_all]{devtools::load_all()}}.}

\item{bundle_opts}{List of arguments passed to \code{\link[=bundle]{bundle()}}.}
}
\description{
Bundles then documents.

Bundles then loads all.
}
