% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-main.R
\name{hash}
\alias{hash}
\alias{hash_set}
\alias{hash_get}
\alias{hash_has}
\alias{hash_match}
\alias{hash_name}
\title{Hash}
\usage{
hash_set(file, type = c("src", "dst"))

hash_get(file, type = c("src", "dst"))

hash_has(file, type = c("src", "dst"))

hash_match(file, type = c("src", "dst"), default = TRUE)

hash_name(file, type = c("src", "dst"))

hash(file)
}
\arguments{
\item{file}{File to hash.}

\item{type}{Type of file to hash.}
}
\description{
Hash

Hash file content
}
\keyword{internal}
