% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools-load--all.R
\name{load_all}
\alias{load_all}
\title{Bundle then loads all}
\usage{
load_all(..., bundle_opts = list())
}
\arguments{
\item{...}{Arguments passed to \code{\link[devtools:load_all]{devtools::load_all()}}.}

\item{bundle_opts}{List of arguments passed to \code{\link[=bundle]{bundle()}}.}
}
\description{
Bundles then loads all.
}
