% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{make_destination_file}
\alias{make_destination_file}
\title{Make Destination File}
\usage{
make_destination_file(srcfile, destdir)
}
\arguments{
\item{srcfile}{Source file to copy.}

\item{destdir}{Desitnation directory.}
}
\description{
Make destination path for file copy.
}
\keyword{internal}
