% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track-tidy.R
\name{tidy}
\alias{tidy}
\title{Tidy}
\usage{
tidy(src = "srcpkg", dest = "R")
}
\arguments{
\item{src}{Source directory.}

\item{dest}{Destination directory.}
}
\description{
Tidy copied files, removes all files.
}
