% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roclet-document.R
\name{document}
\alias{document}
\title{Bundle then document}
\usage{
document(..., bundle_opts = list())
}
\arguments{
\item{...}{Arguments passed to \code{\link[devtools:document]{devtools::document()}}.}

\item{bundle_opts}{List of arguments passed to \code{\link[=bundle]{bundle()}}.}
}
\description{
Bundles then documents.
}
